% Generated by roxygen2 (4.0.1): do not edit by hand
\name{as.logic}
\alias{as.logic}
\alias{as.logicYN}
\title{Convert character to logical}
\usage{
as.logic(x, true = TRUE, false = FALSE)

as.logicYN(x)
}
\arguments{
\item{x}{Variable to convert to logical}

\item{true}{Value(s) mapped to \code{TRUE}}

\item{false}{Value(s) mapped to \code{FALSE}}
}
\value{
\code{logical}
}
\description{
Converts vector to logical based on user specified parameters for the TRUE
and false values
}
\details{
Values not identified by \code{true} or \code{false} become \code{NA}
}
\examples{
as.logic( iris$Species, true="setosa", false="virginica" )
yn <- c( "Yes", "NO", "Y", "n" )
as.logicYN(yn)
}
\seealso{
\code{\link{as.logical}}, \code{\link{logical}}, \code{\link{NA}}
}

