% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.mutex.R
\name{as.mutex}
\alias{as.mutex}
\title{Conert mutually exclusive  columns to a factor}
\usage{
as.mutex(data, names, ordered = FALSE)
}
\arguments{
\item{data}{data to convert}

\item{names}{names of columns in \code{data} to treat as mutually exclusive}

\item{ordered}{logical; whether the factor should be ordered}
}
\value{
factor
}
\description{
Conert mutually exclusive  columns to a factor
}
\note{
\itemize{
\item This performs the opposite function of the dummies package. \cr
\item Ensure the variables are integer or logical or can be coerced to on
\item Belong in dummies package
}
}
\examples{
  # as.mutex( ci2, c('saltsnackyuk', 'saltsnackok', 'saltsnackyum' ), ordered=TRUE )
  

}
