% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_chr.R
\name{as_chr}
\alias{as_chr}
\alias{as_chr.default}
\alias{as_chr.logical}
\title{Convert to character}
\usage{
as_chr(x, ..., na)

\method{as_chr}{default}(x, ..., na)

\method{as_chr}{logical}(x, true = "TRUE", false = "FALSE", na)
}
\arguments{
\item{x}{logical vector}

\item{na}{string; value to set for \code{NA} (missing values)}

\item{true}{string for TRUE value}

\item{false}{string for FALSE value}
}
\description{
Convert to character
}
\details{
\code{as_chr} converts \code{x} to character.  It is smarter versions of
\code{\link[base:as.character]{base::as.character()}}.

The default method wraps \code{\link[base:as.character]{base::as.character()}}.

The method for \code{logical} values allows for mapping \code{TRUE}, \code{FALSE} and
\code{NA} values to
values to  \code{Using non-default values for}true\code{or}false` results in non-invertible
transformation. (true="T", false="F" is also invertible )

This also allows for setting \code{TRUE} or \code{FALSE} values to \code{NA}
}
\examples{
  v <- c(TRUE, FALSE, NA )
  as_chr(v) 
  as_chr(v, "T", "F" )
  as_chr(v, true="T", false=NA )

}
\seealso{
\itemize{
\item \code{\link[base:as.character]{base::as.character()}} which is a primitive method for which
S3 classes cannot be defined.
\item \href{https://cran.r-project.org/web/packages/na.tools/index.html}{na.tools}
package for functions that replace and impute \code{NA} in vectors.
}
}
