% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/can.R
\name{can}
\alias{can}
\title{Determine if the a vector can be coerced into the given clas}
\usage{
can(object, class)
}
\arguments{
\item{object}{object;}

\item{class}{character; name of class to check if it can be coerced}
}
\value{
logical
}
\description{
Determine if the a vector can be coerced into the given clas
}
\examples{
  can( 1.0, 'integer' )
  can( "foo", 'integer' )
      
}
\seealso{
\code{\link[methods]{canCoerce}}
  \code{\link[methods]{setAs}}
}

