% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coerce_each.R
\name{coerce_each}
\alias{coerce_each}
\alias{coerce_each.data.frame}
\alias{coerce_each.data.table}
\title{Coerce each elements of a container class from one class to a different
class}
\usage{
coerce_each(x, from, to)

\method{coerce_each}{data.frame}(x, from, to)

\method{coerce_each}{data.table}(x, from, to)
}
\arguments{
\item{x}{A container object such as a \code{list}, \code{data.frame} or
\code{data.table}}

\item{from}{A character vector containing the classes to coerce from}

\item{to}{A character vector containing}
}
\value{
An object with the same class as \code{x}

\code{data.table}

\code{data.table}
}
\description{
Coerce each elements of a container class from one class to a different
class
}
\examples{
  coerce_each( iris,  "factor", "character" ) 
}
