% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coerce_integer.R
\name{coerce_integer}
\alias{coerce_integer}
\title{Coerce to Integer}
\usage{
coerce_integer(x, strip_units = TRUE)
}
\arguments{
\item{x}{values to be coerced to numeric}

\item{strip_units}{boolean. Whether to strip the unit from the results.}
}
\value{
A numeric field
}
\description{
Coerces a value to numeric by removing any leading non-digit, non-decimal 
fields occuring at the end of the character.
}
\examples{
  coerce_integer( '32 ft/s^2' )
  coerce_integer('1,001 nights' )
  
}
\seealso{
The stringr package.
}
