% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dry_run.R
\name{dry_run}
\alias{dry_run}
\title{dry_run}
\usage{
dry_run(x, fun, type = c("all"))
}
\arguments{
\item{x}{object}

\item{fun}{function}
}
\value{
\code{x} reporting with a message about changes to those values
}
\description{
Like the identity function but reports on changes in values caused by the
application of the function.
}
\examples{
  c(-2:2) \%>\% dry_run( sqrt ) 
  c(-2:2) \%>\% dry_run( sqrt, "all" )
  c(-2:2) \%>\% dry_run( sqrt, "changed" )
  c(-2:2) \%>\% dry_run( sqrt, "unchanged" )   


  data(iris)
  setDT(iris)
  iris[ , Species := Species \%>\% dry_run( toupper ) ] 
  
  iris$Species # unchanged
  
}
