% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_excel_date.R
\docType{data}
\name{excel_origin}
\alias{excel_origin}
\title{excel_origin}
\format{An object of class \code{Date} of length 1.}
\usage{
excel_origin
}
\description{
Date origin according to Microsoft Excel
}
\keyword{datasets}

