% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_age.R
\name{p_age}
\alias{p_age}
\title{parse age}
\usage{
p_age(x)
}
\arguments{
\item{x}{character; age represented as  numeric with a Y, M, D suffic}
}
\value{
numeric age in years
}
\description{
parse age formatted as 34Y, 12M, 20D, etc. to an age in years
}
\details{
Parses a character string to yield an age in years.
}
\examples{
  
  c('34Y', '11M', '12D') \%>\% p_age           

}

