% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_currency.R
\name{p_currency}
\alias{p_currency}
\title{Parse US Currency from text field}
\usage{
p_currency(x)
}
\arguments{
\item{x}{character; text currency values

Parses a character represented currency value to numeric value}
}
\description{
Parse US Currency from text field
}
\examples{
  factor("$1,000.20") \%>\% from_currency

}

