% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_dollar.R
\name{p_dollar}
\alias{p_dollar}
\title{Parse US Currency from text field}
\usage{
p_dollar(x)
}
\arguments{
\item{x}{character; text currency values

Parses a character represented numeric USD value by removing \code{$} and
\code{","} values.}
}
\description{
Parse US Currency from text field
}
\examples{
  factor("$1,000.20") \%>\% from_currency

}
\seealso{
\code{scales::dollar} which is the inverse operations that expresses a
numeric value as a scalar
}
