% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_excel_date.R
\name{p_excel_dt}
\alias{p_excel_dt}
\title{Parse Excel Numeric Date/Time}
\usage{
p_excel_dt(x)
}
\arguments{
\item{x}{numeric; vector of Excel date/times expressed as numeric}
}
\value{
as.POSIX_ct
}
\description{
Parse Excel Numeric Date/Time
}
\details{
At times readxl::read_excel does not get the days right
}
\seealso{
\code{\link[readxl]{read_excel}}
}

