% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_percent.R
\name{p_percent}
\alias{p_percent}
\title{Parse percent value}
\usage{
p_percent(x)
}
\arguments{
\item{x}{character; a value that can be coerced to a character that has a 
percentage sign at the end}
}
\description{
Parse percent value
}
\details{
Returns numeric value with the percentage sign stripped
}

