% Generated by roxygen2 (4.0.1): do not edit by hand
\name{strip_units}
\alias{strip_units}
\title{Strip unit-like characters from a vector}
\usage{
strip_units(x)
}
\arguments{
\item{x}{vector, usually character.}
}
\value{
A vector of the same class as \code{x}, but with the trailing
non-number elements removed. It does not perform a coercion.
}
\description{
Strip everything after the first set of number like items
}
\note{
consider adding an \code{as} parameter to perform coercions. See 2nd
example for comma left in.
}
\examples{
strip_units( '32 ft/s^2')
  strip_units( '1,001 nights' )
}

