% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tf_try.R
\name{tf_try}
\alias{tf_try}
\alias{is.tf_try}
\alias{print.tf_try}
\alias{tf_end}
\title{Try multiple transformations}
\usage{
tf_try(x, fun)

is.tf_try(x)

\method{print}{tf_try}(x, ...)

tf_end(x)
}
\arguments{
\item{x}{vector}

\item{fun}{}
}
\description{
Attempt multiple transformations of a vector
}
\details{
\code{tf_try} transforms the data but keep track of the original vector
allowing for multiple attempts at transforming the vectors\preformatted{ x \%>\% 
   tf_try( ... ) \%>\% 
   tf_try( .... ) \%>\% 
   tf_end()             # optional 
}

This is very handy for transforming data that allowed or didn't enforce a
particular format such as a date.
}
\examples{
  
 # Contrived example sqrt(-x) 
  -2:2 \%>\%  
    tf_try( sqrt ) \%>\% 
    tf_try( . \%>\% abs \%>\% sqrt )
  
 # Shift x  
  -2:2 \%>\%  
    tf_try( sqrt ) \%>\% 
    tf_try( function(x) sqrt( x+ abs(min(x)) ) ) 
 
  c(-2,1,0,NA,1,2) \%>\% 
    tf_try( sqrt ) \%>\% 
    tf_try( function(x) sqrt( x+ abs(min(x)) ) ) 
     
}
