% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/try_as.R
\name{try_as}
\alias{try_as}
\title{try_as}
\usage{
try_as(object, class)
}
\arguments{
\item{object}{object}

\item{class}{string; name of class}
}
\value{
a object of the new class or \code{NULL} if as produces an error
}
\description{
try_as
}
\details{
\code{try_as} attempts to convert \code{object} to \code{class}.

warnings are surpressed; errors return NULL
}
\examples{

  nums <- seq(1.0,2.0,0.2)
  ints <- 1:5
  chrs <- letters[1:5]
  dts  <- seq( Sys.Date(), length.out=5, by="1 week") 
  posx <- as.POSIXct( dts )
 
  cls <- c('numeric','integer','character','Date','POSIXct')
  
  nums \%>\% try_as("numeric")
  nums \%>\% try_as("integer")
  nums \%>\% try_as("character")
  nums \%>\% try_as("Date")
  nums \%>\% try_as("POSIXct") 
    
  ints \%>\% try_as("numeric")
  ints \%>\% try_as("integer")
  ints \%>\% try_as("character")
  ints \%>\% try_as("Date")
  ints \%>\% try_as("POSIXct") 

  chrs \%>\% try_as("numeric")
  chrs \%>\% try_as("integer")
  chrs \%>\% try_as("character")
  chrs \%>\% try_as("Date")
  chrs \%>\% try_as("POSIXct") 

  dts \%>\% try_as("numeric")
  dts \%>\% try_as("integer")
  dts \%>\% try_as("character")
  dts \%>\% try_as("Date")
  dts \%>\% try_as("POSIXct") 
 
  posx \%>\% try_as("numeric")
  posx \%>\% try_as("integer")
  posx \%>\% try_as("character")
  posx \%>\% try_as("Date")
  posx \%>\% try_as("POSIXct") 
  
}
\seealso{
\itemize{
\item \code{\link[=can_be]{can_be()}}
}
}
