% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/as.logic.R
\name{as.logic}
\alias{as.logic}
\alias{as.logicYN}
\title{Convert character to logical}
\usage{
as.logic(x, true = TRUE, false = FALSE)

as.logicYN(x)
}
\arguments{
\item{x}{vector  to convert to logical}

\item{true}{Value(s) mapped to \code{TRUE}}

\item{false}{Value(s) mapped to \code{FALSE}}
}
\value{
\code{logical} vector version of \code{x}.
}
\description{
Converts vector to logical based on user specified parameters for the TRUE
and false values
}
\details{
\code{as.logic} coerces \code{x} into a logical vector in which the values
of specified by \code{true} become \code{TRUE}. etc. Values not identified
by \code{true} or \code{false} become \code{NA}
}
\examples{
as.logic( iris$Species, true="setosa", false="virginica" )
yn <- c( "Yes", "NO", "Y", "n" )
  as.logicYN(yn)
}
\seealso{
\code{\link[base]{as.logical}} \cr
  \code{\link[base]{logical}} \cr
  \code{\link[base]{NA}} \cr
}

