% Generated by roxygen2 (4.0.1): do not edit by hand
\name{as.mutex}
\alias{as.mutex}
\title{Conert mutually exclusive  columns to a factor}
\usage{
as.mutex(data, names, ordered = FALSE)
}
\arguments{
\item{data}{data to convert}

\item{names}{names of columns in \code{data} to treat as mutually exclusive}

\item{ordered}{logical; whether the factor should be ordered}
}
\value{
factor
}
\description{
Conert mutually exclusive  columns to a factor
}
\note{
- This performs the opposite function of the dummies package. \cr
  - Ensure the variables are integer or logical or can be coerced to on
}
\examples{
# as.mutex( ci2, c('saltsnackyuk', 'saltsnackok', 'saltsnackyum' ), ordered=TRUE )
}

