% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/coercion-package.R
\docType{package}
\name{coercion package}
\alias{coercion package}
\alias{coercion package-package}
\title{coercion package}
\description{
coercion package
}
\details{
The coercion package contains utilities for coercing between various
classes, types and modes.
}
\examples{
data(iris)
  coerce_each( iris, 'character' )
}

