% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_excel_date.R
\name{p_excel_dt}
\alias{p_excel_dt}
\title{Parse Excel Numeric Date/Time}
\usage{
p_excel_dt(x)
}
\arguments{
\item{x}{numeric; vector of Excel date/times expressed as numeric}
}
\value{
A Date object if \code{x} is integer; a POSIXct object if \code{x} is 
  numeric. \cr
}
\description{
Parse Excel Numeric Date/Time
}
\details{
At times readxl::read_excel does not get the date and times right and imports
them as numeric instead. This converts those numeric date/times to proper
Date or POSIXct objects.
}
\examples{
  p_excel_dt( 40000 )
  p_excel_dt( c(40000,NA)) 
  p_excel_dt( Sys.Date() )
  p_excel_dt( Sys.time() )   
  p_excel_dt( c(Sus)) 
}
\seealso{
\code{\link[readxl]{read_excel}}
}

