% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tf_check.R
\name{tf_check}
\alias{tf_check}
\title{Check transform}
\usage{
tf_check(data, expr, .all = FALSE)
}
\arguments{
\item{data}{table of data to function as environment for \code{expr}}

\item{expr}{expression to evaluate}

\item{.all}{logical whether to display all the observation or
just the failed transformations. (Default: `FALSE``)}
}
\value{
A data.table with the result of the expression followed by the
names used from the columns
\itemize{
\item unexpected where one or more of the args are na
\item any missing
\item failures : no missing args,
}
}
\description{
Check transformation values
}
\examples{

  data(iris) 
  setDT(iris)
  iris[ , Species := Species \%>\% as.character() ]

  iris \%>\% tf_check( Species \%>\% as.character() )
  iris \%>\% tf_check( Species \%>\% as.numeric()  )

}
